# Conflicts:
#	.gitmodules
#	base/message_loop/message_pump_apple.mm
#	base/message_loop/message_pump_kqueue.cc
#	chrome/services/mac_notifications/mac_notification_service_un_unittest.mm
#	ui/views/style/typography_provider.cc

+++ android_webview/java/src/org/chromium/android_webview/AwBrowserContext.java
@@ -83,16 +83,17 @@ public class AwBrowserContext implements BrowserContextHandle {
         // Register MemoryPressureMonitor callbacks and make sure it polls only if there is at
         // least one WebView around.
         MemoryPressureMonitor.INSTANCE.registerComponentCallbacks();
-        AwContentsLifecycleNotifier.addObserver(new AwContentsLifecycleNotifier.Observer() {
-            @Override
-            public void onFirstWebViewCreated() {
-                MemoryPressureMonitor.INSTANCE.enablePolling();
-            }
-            @Override
-            public void onLastWebViewDestroyed() {
-                MemoryPressureMonitor.INSTANCE.disablePolling();
-            }
-        });
+        AwContentsLifecycleNotifier.getInstance().addObserver(
+                new AwContentsLifecycleNotifier.Observer() {
+                    @Override
+                    public void onFirstWebViewCreated() {
+                        MemoryPressureMonitor.INSTANCE.enablePolling();
+                    }
+                    @Override
+                    public void onLastWebViewDestroyed() {
+                        MemoryPressureMonitor.INSTANCE.disablePolling();
+                    }
+                });
     }
 
     @VisibleForTesting
@@ -232,10 +233,15 @@ public class AwBrowserContext implements BrowserContextHandle {
      * <p>
      * Name must be non-null and valid Unicode.
      *
-     * @throws IllegalStateException if trying to delete the default profile or a profile which is
-     *                               in use.
+     * @throws IllegalArgumentException if trying to delete the default profile.
+     * @throws IllegalStateException if trying to delete a profile which is in use.
      */
-    public static boolean deleteNamedContext(String name) {
+    public static boolean deleteNamedContext(String name)
+            throws IllegalArgumentException, IllegalStateException {
+        final String defaultContextName = AwBrowserContextJni.get().getDefaultContextName();
+        if (name.equals(defaultContextName)) {
+            throw new IllegalArgumentException("Cannot delete the default profile");
+        }
         return AwBrowserContextJni.get().deleteNamedContext(name);
     }
